Qutebrowser Dracula Theme
Up to now I've maintained a local copy of dracula's qutebrowser theme,
instead of using a submodule. This is due to the use of a symlink to put
the theme files into XDG_CONFIG/qutebrowser; submodules can't be
symlinked using home-manager due to the submodule files not being added
to the nix store. To fix this:

- I've added qutebrowser/dracula theme as a flake input
- Referenced the outPath attribute of that input in the symlink.

The theme hasnt been updated in over 2 years but I dont like having
python code in here unless it needs to be. Now its a flake input.